Move pathy to separate file
